Move dhcp-discover into a dedicated "tools" target
